+++ src/factories/index.tsx
@@ -1,1 +1,2 @@
 export * from "./ReduxActionConstants";
+export * from "./Service";