+++ src/services/index.tsx
@@ -1,1 +1,2 @@
 export * from "./buildClient";
+export * from "./BuildClientWithCanceler";