+++ CHANGELOG.md
@@ -6,6 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2020-07-12
+### Added
+- `buildClient` - onResponseFulfilled, onRequestFulfilled, onRequestRejected callbacks in options
+- `BulidClientWithCanceler` - constructor, similar to buildClient, but with request canceling builtin capability 
+- `useFetch` - addition of options, onCancelMsg (`string`) & onCancel (`callback`)
+
+### Changed
+- `buildClient` - onResponseError to onResponseRejected
+- `useFetch` - internally handles request cancellation if dependency changes (or) component unmounts before api hit is completed
+
+### removed
+- `buildClient` - getToken option
+
 ## [0.1.0] - 2020-07-6
 ### Added
 - tests for all the components and hooks, etc.